decoupled group gemms into their own module
