+++ torchtitan/experiments/deepseek_v3/group_gemms.py
@@ -1,0 +1,256 @@
+import torch
+import torch.nn as nn
+
+try:
+    import deep_gemm
+
+    DEEPGEMM_AVAILABLE = True
+except ImportError:
+    DEEPGEMM_AVAILABLE = False
+
+if DEEPGEMM_AVAILABLE:
+    import dsgemm_kernels
+    import dsgemm_utils
+
+try:
+    import torchao
+
+    TORCHAO_FP8_GG_AVAILABLE = True
+    from torchao.prototype.scaled_grouped_mm import _scaled_grouped_mm
+except ImportError:
+    TORCHAO_FP8_GG_AVAILABLE = False
+    # raise NotImplementedError("Missing TorchAO")
+
+
+# Strategy base class for GroupGEMM implementations
+class GroupGEMMStrategy:
+    """Base class for group gemm strategies"""
+
+    def arrange_expert_weights(self, all_weights, submod_name, module):
+        """prepare expert weights, including prescaling"""
+        raise NotImplementedError("Requires arrange_expert_weights method")
+
+    def execute(self, contig_tokens, m_sizes, m_offsets, module):
+        """Execute the group gemm operation
+
+        Args:
+            contig_tokens: The input tokens, arranged contiguously by expert
+            m_sizes: Sizes of each group
+            m_offsets: Offsets of each group
+            module: The MoE module containing weights and parameters
+
+        Returns:
+            The processed tokens
+        """
+        raise NotImplementedError("GroupGEMM strategy must implement execute method")
+
+    @staticmethod
+    def is_available() -> bool:
+        """Check if this strategy is available on the current system"""
+        return True
+
+
+# ========= Implementations ===================
+__all__ = ["TorchFP8GroupGEMM", "DSGroupGEMM"]
+
+
+class TorchFP8GroupGEMM(GroupGEMMStrategy):
+    """Implementation using TorchAO's _scaled_grouped_mm with FP8 rowwise precision"""
+
+    def __init__(
+        self,
+    ):
+        self.activation_function = nn.SiLU()
+
+    def arrange_expert_weights(self, all_weights, submod_name, module):
+        """prep the expert weights for group gemm usage"""
+        return torch.stack(all_weights)
+
+    def execute(self, contig_tokens, m_sizes, m_offsets, module):
+        # Get weights
+        w_gate = module.get_parameter("gate_proj_weight")
+        w_up = module.get_parameter("up_proj_weight")
+        w_down = module.get_parameter("down_proj_weight")
+
+        # Run first two GEMMs (gate and up projections)
+        gate_proj = _scaled_grouped_mm(
+            contig_tokens,
+            w_gate.transpose(-2, -1),
+            m_offsets,
+            out_dtype=torch.bfloat16,
+        )
+        up_proj = _scaled_grouped_mm(
+            contig_tokens,
+            w_up.transpose(-2, -1),
+            m_offsets,
+            out_dtype=torch.bfloat16,
+        )
+
+        # Apply activation
+        hidden_outputs = self.activation_function(gate_proj) * up_proj
+
+        # Run the third GEMM (down projection)
+        hidden_outputs = _scaled_grouped_mm(
+            hidden_outputs,
+            w_down.transpose(-2, -1),
+            m_offsets,
+            out_dtype=torch.bfloat16,
+        )
+
+        return hidden_outputs
+
+    @staticmethod
+    def is_available() -> bool:
+        return TORCHAO_FP8_GG_AVAILABLE
+
+
+class DSGroupGEMM(GroupGEMMStrategy):
+    """Implementation using DeepGEMM with FP8 quantization"""
+
+    def __init__(self):
+        # Cache for quantized tensors to avoid repeated work
+        self.input_cache = {}
+        self.activation_function = nn.SiLU()
+
+    def arrange_expert_weights(self, all_weights, submod_name, module):
+        """prep the expert weights for group gemm usage"""
+        combined_weights = torch.stack(all_weights)
+
+        fp8, scales = dsgemm_utils.prepare_fp8_weight(combined_weights)
+
+        # prescale weights
+        module.register_parameter(
+            f"{submod_name}_fp8",
+            nn.Parameter(
+                fp8,
+            ),
+        )
+
+        module.register_parameter(
+            f"{submod_name}_scales",
+            nn.Parameter(
+                scales,
+            ),
+        )
+
+        return combined_weights
+
+    def execute(self, contig_tokens, m_sizes, m_offsets, module):
+        # Get only valid tokens
+        valid_tokens = contig_tokens[: m_offsets[-1]]
+
+        # Create indices from offsets without CPU-GPU sync
+        m_indices = dsgemm_utils.create_indices_from_offsets_nosync(m_offsets)
+
+        # Get expert weights for all projections
+        gate_proj_weight_fp8 = module.get_parameter("gate_proj_fp8")
+        gate_proj_scales = module.get_parameter("gate_proj_scales")
+        up_proj_weight_fp8 = module.get_parameter("up_proj_fp8")
+        up_proj_scales = module.get_parameter("up_proj_scales")
+        down_proj_weight_fp8 = module.get_parameter("down_proj_fp8")
+        down_proj_scales = module.get_parameter("down_proj_scales")
+
+        # Get dimensions
+        m_actual_tokens = valid_tokens.shape[0]
+        intermediate_size = module.get_parameter("gate_proj_weight").shape[1]
+        hidden_size = module.get_parameter("down_proj_weight").shape[1]
+
+        # Allocate output buffers
+        gate_proj_out = torch.empty(
+            (m_actual_tokens, intermediate_size),
+            device=contig_tokens.device,
+            dtype=contig_tokens.dtype,
+        )
+        up_proj_out = torch.empty_like(gate_proj_out)
+
+        # Allocate output buffer for down projection
+        down_proj_out = torch.empty(
+            (m_actual_tokens, hidden_size),
+            device=contig_tokens.device,
+            dtype=contig_tokens.dtype,
+        )
+
+        """# Prepare or resize output buffers
+        if not hasattr(module, "_buffer_initialized") or not module._buffer_initialized:
+            module.register_buffer(
+                "gate_proj_buffer",
+                torch.empty(
+                    (m_actual_tokens, intermediate_size),
+                    device=contig_tokens.device,
+                    dtype=contig_tokens.dtype,
+                ),
+            )
+            module.register_buffer(
+                "up_proj_buffer", torch.empty_like(module.gate_proj_buffer)
+            )
+            module.register_buffer(
+                "down_proj_buffer",
+                torch.empty(
+                    (m_actual_tokens, hidden_size),
+                    device=contig_tokens.device,
+                    dtype=contig_tokens.dtype,
+                ),
+            )
+            module._buffer_initialized = True
+        else:
+            # Resize buffers if needed
+            if module.gate_proj_buffer.shape[0] < m_actual_tokens:
+                module.gate_proj_buffer = torch.empty(
+                    (m_actual_tokens, intermediate_size),
+                    device=contig_tokens.device,
+                    dtype=contig_tokens.dtype,
+                )
+                module.up_proj_buffer = torch.empty_like(module.gate_proj_buffer)
+                module.down_proj_buffer = torch.empty(
+                    (m_actual_tokens, hidden_size),
+                    device=contig_tokens.device,
+                    dtype=contig_tokens.dtype,
+                )
+        """
+        # Prepare input in FP8 format (shared by gate and up projections)
+        if module.use_triton_quant:
+            gate_up_input = dsgemm_kernels.groupwise_activation_quant(valid_tokens)
+        else:
+            gate_up_input = dsgemm_utils.prepare_fp8_input(valid_tokens)
+
+        # Run first GEMM (gate projection)
+        deep_gemm.m_grouped_gemm_fp8_fp8_bf16_nt_contiguous(
+            gate_up_input,
+            (gate_proj_weight_fp8, gate_proj_scales),
+            gate_proj_out,
+            m_indices,
+        )
+
+        # Run second GEMM (up projection)
+        deep_gemm.m_grouped_gemm_fp8_fp8_bf16_nt_contiguous(
+            gate_up_input,
+            (up_proj_weight_fp8, up_proj_scales),
+            up_proj_out,
+            m_indices,
+        )
+
+        # Apply activation
+        hidden_states = self.activation_function(gate_proj_out) * up_proj_out
+
+        # Run third GEMM (down projection)
+        if module.use_triton_quant:
+            hidden_states_quantized = dsgemm_kernels.groupwise_activation_quant(
+                hidden_states
+            )
+        else:
+            hidden_states_quantized = dsgemm_utils.prepare_fp8_input(hidden_states)
+
+        deep_gemm.m_grouped_gemm_fp8_fp8_bf16_nt_contiguous(
+            hidden_states_quantized,
+            (down_proj_weight_fp8, down_proj_scales),
+            down_proj_out,
+            m_indices,
+        )
+
+        # Copy results back to contig_tokens
+        contig_tokens[: m_offsets[-1]] = down_proj_out
+        return contig_tokens
+
+    @staticmethod
+    def is_available() -> bool:
+        return DEEPGEMM_AVAILABLE
